requirements.txt
streamlit
tensorflow
google-genai
numpy
opencv-python
matplotlib
pillow